Camp8Finished

+++ Entities/Concrete/Car.cs
@@ -13,7 +13,7 @@ public class Car:IEntity
     public int BrandId { get; set; }
     public int ColorId { get; set; }
     public int ModelYear { get; set; }
-    public double DailyPrice { get; set; }
+    public decimal DailyPrice { get; set; }
     public string Name { get; set; }
     public string Description { get; set; }
 }
